Handle child process failure better (#22)
If the target SharpFuzz process dies without being
able to write the `status` back to the pipe, then
libfuzzer-dotnet gets blocked on its read call.

Fix this by closing our other handle to the write
end of the pipe, so that when the child process
exits, the read call will return with an error.

Also print out the child process exit code if this
happens, so that we can see what the error was.

In addition, convert die_sys to die_lasterror and
print out additional error information.